misc: Remove raven-js from master
We track raven-js and release important bugfixes only from the '4.x'
branch, and never from master.

Over time, the code in raven-js in master and in the 4.x branch got out
of sync.

Removing raven-js from master eliminates confusion and cleans up the
workspace when developing the current 5.x series of packages.